Move CraftingListener to ItemsProtectionListener
